+++ src/main/java/net/werdei/zoomglass/client/NoItemException.java
@@ -1,5 +1,0 @@
-package net.werdei.zoomglass.client;
-
-public class NoItemException extends Exception
-{
-}