+++ src/main/java/net/werdei/zoomglass/mixin/AbstractClientPlayerEntityMixin.java
@@ -1,19 +1,19 @@
 package net.werdei.zoomglass.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
 import net.werdei.zoomglass.client.ZoomglassClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(AbstractClientPlayerEntity.class)
 public class AbstractClientPlayerEntityMixin
 {
-    @Inject(method = "getFovMultiplier", at=@At("HEAD"), cancellable = true)
-    public void forceSpyglassFov(CallbackInfoReturnable<Float> cir)
+    @ModifyReturnValue(method = "getFovMultiplier", at=@At("RETURN"))
+    public float forceSpyglassFov(float original, boolean firstPerson)
     {
-        if (ZoomglassClient.isFakeSpyglassActive())
-            cir.setReturnValue(0.1F);
+        if (firstPerson && ZoomglassClient.isFakeSpyglassActive())
+            return 0.1f;
+        return original;
     }
 }