+++ src/main/java/net/werdei/zoomglass/mixin/MinecraftClientMixin.java
@@ -1,18 +1,17 @@
 package net.werdei.zoomglass.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.option.KeyBinding;
 import net.werdei.zoomglass.client.ZoomglassClient;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(MinecraftClient.class)
 public class MinecraftClientMixin
 {
-    @Redirect(method = "handleInputEvents", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/option/KeyBinding;isPressed()Z", ordinal = 2))
-    private boolean dontCancelQuickSpyglass(KeyBinding keyUse)
+    @ModifyExpressionValue(method = "handleInputEvents", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/option/KeyBinding;isPressed()Z", ordinal = 2))
+    private boolean dontCancelQuickSpyglass(boolean original)
     {
-        return keyUse.isPressed() || ZoomglassClient.isQuickSpyglassActive();
+        return original || ZoomglassClient.isQuickSpyglassActive();
     }
 }