+++ CHANGELOG.md
@@ -1,1 +1,3 @@
-* Updated to 1.21.3
+* Using spyglass that's held in offhand is now supported
+* Fixed broken zooming animation of your character in multiplayer
+* When using a fake spyglass (in creative/spectator), a spyglass sound will be played